DFE-1049: move serverside admin to tools area
