+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Controllers/FileControllerTests.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using GovUk.Education.ExploreEducationStatistics.Admin.Areas.Tools.Controllers;
 using GovUk.Education.ExploreEducationStatistics.Admin.Controllers;
 using GovUk.Education.ExploreEducationStatistics.Admin.Models;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces;
