+++ src/GovUk.Education.ExploreEducationStatistics.Admin.Tests/Controllers/FileUploadControllerTests.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using System.IO;
 using System.Linq;
 using System.Threading.Tasks;
+using GovUk.Education.ExploreEducationStatistics.Admin.Areas.Tools.Controllers;
 using GovUk.Education.ExploreEducationStatistics.Admin.Controllers;
 using GovUk.Education.ExploreEducationStatistics.Admin.Services.Interfaces;
 using GovUk.Education.ExploreEducationStatistics.Content.Model;
