+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Areas/Tools/Views/Home/Index.cshtml
@@ -1,0 +1,49 @@
+﻿@{
+    ViewData["Title"] = "Education statistics publisher";
+}
+
+<h1 class="govuk-heading-l">BAU Team Tools</h1>
+
+<h2 class="govuk-heading-m">Fake services</h2>
+
+<div class="govuk-grid-row">
+    <div class="govuk-grid-column-one-third">
+        <h3 class="govuk-heading-s">Files</h3>
+        <ul class="govuk-list">
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="File" asp-action="List">List files</a>
+            </li>
+            <li>
+                <a class="govuk-link" asp-area="Tools" asp-controller="FileUpload" asp-action="Upload">Upload file</a>
+            </li>
+        </ul>
+    </div>
+    <div class="govuk-grid-column-one-third">
+        <h3 class="govuk-heading-s">Release actions</h3>
+        <ul class="govuk-list">
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="Notifications" asp-action="NotifySubscribers">Notify subscribers</a>
+            </li>
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="Publishing" asp-action="PublishReleaseData">Publish release data</a>
+            </li>
+        </ul>
+    </div>
+    <div class="govuk-grid-column-one-third">
+        <h3 class="govuk-heading-s">Fake CMS</h3>
+        <ul class="govuk-list">
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="Themes" asp-action="Index">Themes</a>
+            </li>
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="Topics" asp-action="Index">Topics</a>
+            </li>
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="Publications" asp-action="Index">Publications</a>
+            </li>
+            <li >
+                <a class="govuk-link" asp-area="Tools" asp-controller="Releases" asp-action="Index">Releases</a>
+            </li>
+        </ul>
+    </div>
+</div>
