+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Areas/Tools/Views/Releases/Delete.cshtml
