+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Areas/Tools/Views/Shared/_Layout.cshtml
@@ -56,8 +56,8 @@
                 </a>
             </div>
             <div class="govuk-header__content">
-                <a href="/tools" class="govuk-header__link govuk-header__link--service-name">
-                    EES - Admin Tools
+                <a href="/" class="govuk-header__link govuk-header__link--service-name">
+                    Explore education statistics
                 </a>
                 <partial name="_LoginPartial" />
             </div>
