+++ src/GovUk.Education.ExploreEducationStatistics.Admin/Areas/Tools/Views/Shared/_LoginPartial.cshtml
@@ -1,0 +1,30 @@
+@using System.Security.Principal
+
+@if (User.Identity.IsAuthenticated)
+{
+    var email = User.Claims.FirstOrDefault(x => x.Type == "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/emailaddress")?.Value;
+    var givenname = User.Claims.FirstOrDefault(x => x.Type == "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/givenname")?.Value;
+    var surname = User.Claims.FirstOrDefault(x => x.Type == "http://schemas.xmlsoap.org/ws/2005/05/identity/claims/surname")?.Value;
+
+                
+    // TOOO - temp work around for dfe accounts to get name
+    if (email.ToLower().Contains("education.gov.uk"))
+    {
+        var address = email.Split('@').First();
+        givenname = address.Split('.').First();
+        surname = address.Split('.').Last();
+    }
+    
+    <button type="button" role="button" class="govuk-header__menu-button js-header-toggle" aria-controls="navigation" aria-label="Show or hide Top Level Navigation">Menu</button>
+    <nav>
+        <ul id="navigation" class="govuk-header__navigation " aria-label="Top Level Navigation">
+            <li class="govuk-header__navigation-item">
+                <a class="govuk-header__link">@givenname @surname</a>
+            </li>
+            <li class="govuk-header__navigation-item">
+                <a class="govuk-header__link" asp-area="AzureAD" asp-controller="Account" asp-action="SignOut">Sign out</a>
+            </li>
+        </ul>
+    </nav>
+}
+
