+++ src/GovUk.Education.ExploreEducationStatistics.Admin/GovUk.Education.ExploreEducationStatistics.Admin.csproj
@@ -89,6 +89,41 @@
     <_ContentIncludedByDefault Remove="wwwroot\assets\images\icon-important.png" />
     <_ContentIncludedByDefault Remove="wwwroot\assets\images\icon-pointer-2x.png" />
     <_ContentIncludedByDefault Remove="wwwroot\assets\images\icon-pointer.png" />
+    <_ContentIncludedByDefault Remove="Views\FileUpload\Upload.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\File\List.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Home\Index.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Home\Privacy.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Notifications\NotificationsSent.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Notifications\NotifySubscribers.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publications\Create.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publications\Delete.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publications\Details.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publications\Edit.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publications\Index.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publishing\PublishReleaseData.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Publishing\ReleaseDataPublished.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Releases\Create.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Releases\Delete.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Releases\Details.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Releases\Edit.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Releases\Index.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Shared\Error.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Shared\_CookieConsentPartial.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Shared\_Layout.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Shared\_LoginPartial.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Shared\_ValidationScriptsPartial.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Themes\Create.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Themes\Delete.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Themes\Details.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Themes\Edit.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Themes\Index.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Topics\Create.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Topics\Delete.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Topics\Details.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Topics\Edit.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\Topics\Index.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\_ViewImports.cshtml" />
+    <_ContentIncludedByDefault Remove="Views\_ViewStart.cshtml" />
   </ItemGroup>
