--- src/GovUk.Education.ExploreEducationStatistics.Admin/Views/Home/Index.cshtml
@@ -1,66 +1,0 @@
-﻿@{
-    ViewData["Title"] = "Education statistics publisher";
-}
-
-@if (User.Identity.IsAuthenticated)
-{
-    <h1 class="govuk-heading-l">Dashboard</h1>
-
-    <h2 class="govuk-heading-m">Fake services</h2>
-    
-    <div class="govuk-grid-row">
-        <div class="govuk-grid-column-one-third">
-            <h3 class="govuk-heading-s">Files</h3>
-            <ul class="govuk-list">
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="File" asp-action="List">List files</a>
-                </li>
-                <li>
-                    <a class="govuk-link" asp-area="" asp-controller="FileUpload" asp-action="Upload">Upload file</a>
-                </li>
-            </ul>
-        </div>
-        <div class="govuk-grid-column-one-third">
-            <h3 class="govuk-heading-s">Release actions</h3>
-            <ul class="govuk-list">
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="Notifications" asp-action="NotifySubscribers">Notify subscribers</a>
-                </li>
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="Publishing" asp-action="PublishReleaseData">Publish release data</a>
-                </li>
-            </ul>
-        </div>
-        <div class="govuk-grid-column-one-third">
-            <h3 class="govuk-heading-s">Fake CMS</h3>
-            <ul class="govuk-list">
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="Themes" asp-action="Index">Themes</a>
-                </li>
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="Topics" asp-action="Index">Topics</a>
-                </li>
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="Publications" asp-action="Index">Publications</a>
-                </li>
-                <li >
-                    <a class="govuk-link" asp-area="" asp-controller="Releases" asp-action="Index">Releases</a>
-                </li>
-            </ul>
-        </div>
-    
-    </div>
-
-}
-else
-{
-    <h1 class="govuk-heading-l">
-        <span class="govuk-caption-l">Education statistics publisher</span>
-        Sign-in
-    </h1>
-    <p class="govuk-body">
-        Use this service to publish official Department for Education (DfE) statistics and data for state-funded schools in England.
-    </p>
-
-    <a class="govuk-button govuk-button--start" asp-area="AzureAD" asp-controller="Account" asp-action="SignIn">Sign-in</a>
-}
