--- src/GovUk.Education.ExploreEducationStatistics.Admin/Views/Shared/_LoginPartial.cshtml
@@ -1,17 +1,0 @@
-@using System.Security.Principal
-
-@if (User.Identity.IsAuthenticated)
-{
-    <button type="button" role="button" class="govuk-header__menu-button js-header-toggle" aria-controls="navigation" aria-label="Show or hide Top Level Navigation">Menu</button>
-    <nav>
-        <ul id="navigation" class="govuk-header__navigation " aria-label="Top Level Navigation">
-            <li class="govuk-header__navigation-item">
-                <a class="govuk-header__link">@User.Identity.Name.Split('#').Last()</a>
-            </li>
-            <li class="govuk-header__navigation-item">
-                <a class="govuk-header__link" asp-area="AzureAD" asp-controller="Account" asp-action="SignOut">Sign out</a>
-            </li>
-        </ul>
-    </nav>
-}
-
